Extracted playback controls to new class
